removed proxy related vars and methods from RestService
We should not deal with HttpUrlConnection proxy settings in our RestService.
First reason is that this assumes that we're using HttpUrlConnection. This is
necessarily the case sinve we hide this behind an interface IHttpClient which
may be implemented with whatever http library you want.
Furthermore this is very unexpected since one would expect that whatever he set's
to the system properties is used by whatever HttpUrlConnection in the JVM.
The code that was in place before simply overrode the system properties by the
values handed to the openshift-java-client.